+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 1.2.0 (2020-08-11)
+
+* Add date and time to report folders [Issue 8](https://github.com/grasshopper7/extentreports-cucumber6-adapter/issues/8)
+* Fix report folder name issue [Issue 9](https://github.com/grasshopper7/extentreports-cucumber5-adapter/issues/9)
+
 ## 1.1.0 (2020-07-21)
 
 * Support for Json formatter [Issue 3](https://github.com/grasshopper7/extentreports-cucumber6-adapter/issues/3)